introduce visualisation script
